Skip updating VSR status if it's not changed
Previously, the IC would update the status of a VSR (via an API call
to the k8s API) even if the status of the VSR wasn't changed.
The unnecessary API calls lead to the increased IC pod startup time
for cases when a VS had multiple (hundreds) VSRs.

The commit fixes the problem - the IC will skip the update if the
status of a VSR isn't changed.